ps aux fix

+++ deploy-test.sh
@@ -1,28 +1,33 @@
 #!/usr/bin/env bash
 
-targetHost=yogaapp@test.ino01
+targetHost=$1
 deployName=go-devops
+fast=$2
 
-rm -fr $deployName.linux.bin $deployName.linux.bin.bz2
-./gobin.sh
-env GOOS=linux GOARCH=amd64 go build -o $deployName.linux.bin
-bzip2 $deployName.linux.bin
-rsync -avz --human-readable --progress -e "ssh -p 22" ./$deployName.linux.bin.bz2 $targetHost:.
+if [ "$fast" == "fast" ]; then
+    echo "jump building in fast mode"
+else
+    echo "rebuilding"
+    rm -fr $deployName.linux.bin $deployName.linux.bin.bz2
+    ./gobin.sh
+    env GOOS=linux GOARCH=amd64 go build -o $deployName.linux.bin
+    upx $deployName.linux.bin
+fi
+
+
+rsync -avz --human-readable --progress -e "ssh -p 22" ./$deployName.linux.bin $targetHost:.
 #scp ./$deployName.linux.bin.bz2 $targetHost:./
 scp ./deploy-agent.sh $targetHost:.
 
 ssh -tt $targetHost "bash -s" << eeooff
 mkdir -p ./app/$deployName/
 cd ./app/$deployName/
-mv ~/$deployName.linux.bin.bz2 .
+mv -f ~/$deployName.linux.bin .
 mv ~/deploy-agent.sh .
 ps -ef|grep $deployName|grep -v grep|awk '{print \$2}'|xargs -r kill -9
 chmod +x ./deploy-agent.sh
 ./deploy-agent.sh $deployName app01 app/$deployName
 ./deploy-agent.sh $deployName cpapp@app01 app/$deployName 6889 false
-./deploy-agent.sh $deployName app02 app/$deployName
-rm -f $deployName.linux.bin
-bzip2 -d $deployName.linux.bin.bz2
 nohup ./$deployName.linux.bin 2>&1 >> nohup.out &
 exit
 eeooff
